Debug and new tests: - Fix HalfMove memory leak - Fix integer overflow - Update combined tests

+++ src/PGN.cpp
@@ -45,7 +45,7 @@ void PGN::ParseNextGame() {
   if (IS_EOF) {
     throw NoGameFound();
   }
-  int loc = NextNonBlank(LastGameEndLoc);
+  long loc = NextNonBlank(LastGameEndLoc);
   if (IS_EOF) {
     throw NoGameFound();
   }
